feat: Use explicit lifetime for RpcFunc, thus allowing non-static lifetimes
This change is not effective if you spawn `evt_loop` with `tokio::spawn`
as the latter requires `'static` lifetime, and it poisones the client,
and the core with `'static` requirement. I ended up using `Arc` for the
context state anyway, so these changes are not solving any particular
issue, but I still think it might be worth having it specified this way
to potentially relax some future use-cases.